requirements.txt
kivy
kivymd
gTTS
playsound
SpeechRecognition
pyttsx3
python-docx